chore: refactor long files into a few shorter ones
