+++ libs/application/templates/accident-notification/src/forms/AccidentNotificationForm/aboutTheAccidentSection/accidentTypeScreen.ts
@@ -1,0 +1,17 @@
+import { buildRadioField, buildSubSection } from '@island.is/application/core'
+import { accidentType } from '../../../lib/messages'
+import { getAccidentTypeOptions } from '../../../utils'
+
+export const accidentTypeScreen = buildSubSection({
+  id: 'accidentType.section',
+  title: accidentType.general.subsectionTitle,
+  children: [
+    buildRadioField({
+      id: 'accidentType.radioButton',
+      width: 'half',
+      title: accidentType.general.heading,
+      description: accidentType.general.description,
+      options: (formValue) => getAccidentTypeOptions(formValue.answers),
+    }),
+  ],
+})
