+++ libs/application/templates/accident-notification/src/forms/AccidentNotificationForm/externalDataSection/index.ts
@@ -1,0 +1,9 @@
+import { buildSection } from '@island.is/application/core'
+import { agreementDescriptionScreen } from './agreementDescriptionScreen'
+import { accidentNotificationScreen } from './accidentNotificationScreen'
+
+export const externalDataSection = buildSection({
+  id: 'ExternalDataSection',
+  title: 'Meðferð á gögnum',
+  children: [agreementDescriptionScreen, accidentNotificationScreen],
+})
