+++ libs/application/templates/accident-notification/src/forms/AccidentNotificationForm/index.ts
@@ -3,13 +3,14 @@ import { Form, FormModes } from '@island.is/application/types'
 import Logo from '../../assets/Logo'
 import { application } from '../../lib/messages'
 
-import { aboutTheAccidentSection } from './aboutTheAccidentSection'
 import { conclusionSection } from './conclusionSection'
-import { externalDataSection } from './externalDataSection'
+
 import { overviewSection } from './overviewSection'
-import { whoIsTheNotificationForSection } from './whoIsTheNotificationForSection'
 import { betaTestSection } from './betaTestSection'
 import { applicantInformationSection } from './applicantInformationSection'
+import { whoIsTheNotificationForSection } from './whoIsTheNotificationForSection'
+import { externalDataSection } from './externalDataSection'
+import { aboutTheAccidentSection } from './aboutTheAccidentSection'
 
 export const AccidentNotificationForm: Form = buildForm({
   id: 'AccidentNotificationForm',
